OneStopNFTCreator/requirements.txt
py-algorand-sdk==2.1.0
beaker-pyteal==0.5.4